Convert ContextualMenu and ContextualMenuItem to CSS in JS
