+++ packages/office-ui-fabric-react/src/components/ContextualMenu/ContextualMenu.base.tsx
@@ -1,14 +1,18 @@
 import * as React from 'react';
 import {
   IContextualMenuProps,
-  IContextualMenuItem,
   ContextualMenuItemType,
   IContextualMenuStyleProps,
   IContextualMenuStyles
 } from './ContextualMenu.types';
+import {
+  IContextualMenuItemStyles,
+  IContextualMenuItem,
+  IContextualMenuItemStyleProps
+} from './ContextualMenuItem.types';
 import { DirectionalHint } from '../../common/DirectionalHint';
 import { FocusZone, FocusZoneDirection, IFocusZoneProps, FocusZoneTabbableElements } from '../../FocusZone';
-import { IMenuItemClassNames, getItemClassNames } from './ContextualMenu.classNames';
+import { getStyles as getItemStyles } from './ContextualMenuItem.styles';
 import {
   assign,
   BaseComponent,
@@ -20,7 +24,6 @@ import {
   getLastFocusable,
   getRTL,
   getWindow,
-  IClassNames,
   IPoint,
   KeyCodes,
   shouldWrapFocus
@@ -35,6 +38,7 @@ import {
   ContextualMenuButton,
   ContextualMenuAnchor
 } from './ContextualMenuItemWrapper/index';
+import { IProcessedStyleSet } from '@uifabric/styling/lib/MergeStyles';
 
 const getClassNames = classNamesFunction<IContextualMenuStyleProps, IContextualMenuStyles>();
 
@@ -75,6 +79,8 @@ export function canAnyMenuItemsCheck(items: IContextualMenuItem[]): boolean {
 
 const NavigationIdleDelay = 250 /* ms */;
 
+const getItemClassNames = classNamesFunction<IContextualMenuItemStyleProps, IContextualMenuItemStyles>();
+
 @withResponsiveMode
 export class ContextualMenuBase extends BaseComponent<IContextualMenuProps, IContextualMenuState> {
   // The default ContextualMenu properties have no items and beak, the default submenu direction is right and top.
@@ -98,7 +104,7 @@ export class ContextualMenuBase extends BaseComponent<IContextualMenuProps, ICon
 
   private _adjustedFocusZoneProps: IFocusZoneProps;
 
-  private _classNames: IClassNames<IContextualMenuStyles>;
+  private _classNames: IProcessedStyleSet<IContextualMenuStyles>;
 
   constructor(props: IContextualMenuProps) {
     super(props);
@@ -382,20 +388,37 @@ export class ContextualMenuBase extends BaseComponent<IContextualMenuProps, ICon
     // We only send a dividerClassName when the item to be rendered is a divider. For all other cases, the default divider style is used.
     const dividerClassName = item.itemType === ContextualMenuItemType.Divider ? item.className : undefined;
     const subMenuIconClassName = item.submenuIconProps ? item.submenuIconProps.className : '';
-    const getTheItemClassNames = item.getItemClassNames || getItemClassNames;
-    const itemClassNames = getTheItemClassNames(
-      this.props.theme!,
-      isItemDisabled(item),
-      this.state.expandedMenuItemKey === item.key,
-      !!getIsChecked(item),
-      !!item.href,
-      iconProps.iconName !== 'None',
-      item.className,
-      dividerClassName,
-      iconProps.className,
-      subMenuIconClassName,
-      item.primaryDisabled
-    );
+
+    let itemClassNames;
+    if (item.getItemClassNames) {
+      itemClassNames = item.getItemClassNames(
+        this.props.theme!,
+        isItemDisabled(item),
+        this.state.expandedMenuItemKey === item.key,
+        !!getIsChecked(item),
+        !!item.href,
+        iconProps.iconName !== 'None',
+        item.className,
+        dividerClassName,
+        iconProps.className,
+        subMenuIconClassName,
+        item.primaryDisabled
+      );
+    } else {
+      itemClassNames = getItemClassNames(getItemStyles, {
+        theme: this.props.theme!,
+        disabled: isItemDisabled(item),
+        expanded: this.state.expandedMenuItemKey === item.key,
+        checked: !!getIsChecked(item),
+        isAnchorLink: !!item.href,
+        knownIcon: iconProps.iconName !== 'None',
+        itemClassName: item.className,
+        dividerClassName,
+        iconClassName: iconProps.className,
+        subMenuClassName: subMenuIconClassName,
+        primaryDisabled: item.primaryDisabled
+      });
+    }
 
     if (item.text === '-' || item.name === '-') {
       item.itemType = ContextualMenuItemType.Divider;
@@ -431,7 +454,7 @@ export class ContextualMenuBase extends BaseComponent<IContextualMenuProps, ICon
 
   private _renderSectionItem(
     item: IContextualMenuItem,
-    menuClassNames: IMenuItemClassNames,
+    menuClassNames: IProcessedStyleSet<IContextualMenuItemStyles>,
     index: number,
     hasCheckmarks: boolean,
     hasIcons: boolean
@@ -479,7 +502,7 @@ export class ContextualMenuBase extends BaseComponent<IContextualMenuProps, ICon
   private _renderListItem(
     content: React.ReactNode,
     key: string | number,
-    classNames: IMenuItemClassNames,
+    classNames: IProcessedStyleSet<IContextualMenuItemStyles>,
     title?: string
   ) {
     return (
@@ -491,7 +514,7 @@ export class ContextualMenuBase extends BaseComponent<IContextualMenuProps, ICon
 
   private _renderSeparator(
     index: number,
-    classNames: IMenuItemClassNames,
+    classNames: IProcessedStyleSet<IContextualMenuItemStyles>,
     top?: boolean,
     fromSection?: boolean
   ): React.ReactNode {
@@ -509,7 +532,7 @@ export class ContextualMenuBase extends BaseComponent<IContextualMenuProps, ICon
 
   private _renderNormalItem(
     item: IContextualMenuItem,
-    classNames: IMenuItemClassNames,
+    classNames: IProcessedStyleSet<IContextualMenuItemStyles>,
     index: number,
     focusableElementIndex: number,
     totalItemCount: number,
@@ -561,7 +584,7 @@ export class ContextualMenuBase extends BaseComponent<IContextualMenuProps, ICon
 
   private _renderHeaderMenuItem(
     item: IContextualMenuItem,
-    classNames: IMenuItemClassNames,
+    classNames: IProcessedStyleSet<IContextualMenuItemStyles>,
     index: number,
     hasCheckmarks: boolean,
     hasIcons: boolean
@@ -583,7 +606,7 @@ export class ContextualMenuBase extends BaseComponent<IContextualMenuProps, ICon
 
   private _renderAnchorMenuItem(
     item: IContextualMenuItem,
-    classNames: IMenuItemClassNames,
+    classNames: IProcessedStyleSet<IContextualMenuItemStyles>,
     index: number,
     focusableElementIndex: number,
     totalItemCount: number,
@@ -620,7 +643,7 @@ export class ContextualMenuBase extends BaseComponent<IContextualMenuProps, ICon
 
   private _renderButtonItem(
     item: IContextualMenuItem,
-    classNames: IMenuItemClassNames,
+    classNames: IProcessedStyleSet<IContextualMenuItemStyles>,
     index: number,
     focusableElementIndex: number,
     totalItemCount: number,
@@ -659,7 +682,7 @@ export class ContextualMenuBase extends BaseComponent<IContextualMenuProps, ICon
 
   private _renderSplitButton(
     item: IContextualMenuItem,
-    classNames: IMenuItemClassNames,
+    classNames: IProcessedStyleSet<IContextualMenuItemStyles>,
     index: number,
     focusableElementIndex: number,
     totalItemCount: number,
