+++ packages/office-ui-fabric-react/src/components/ContextualMenu/ContextualMenu.styles.ts
@@ -1,5 +1,6 @@
 import { IContextualMenuStyleProps, IContextualMenuStyles } from './ContextualMenu.types';
 import { getGlobalClassNames, FontWeights } from '../../Styling';
+import { ContextualMenuItemHeight } from 'office-ui-fabric-react/lib/components/ContextualMenu/ContextualMenuItem.styles';
 
 const GlobalClassNames = {
   root: 'ms-ContextualMenu',
@@ -15,8 +16,6 @@ export const getStyles = (props: IContextualMenuStyleProps): IContextualMenuStyl
   const classNames = getGlobalClassNames(GlobalClassNames, theme);
 
   const { palette, fonts, semanticColors } = theme;
-
-  const ContextualMenuItemHeight = '32px';
 
   return {
     root: [
