+++ packages/office-ui-fabric-react/src/components/ContextualMenu/ContextualMenuItem.base.tsx
@@ -2,72 +2,24 @@ import * as React from 'react';
 import { hasSubmenu, getIsChecked } from '../../utilities/contextualMenu/index';
 import { BaseComponent, getRTL } from '../../Utilities';
 import { Icon } from '../../Icon';
-import { IContextualMenuItemProps } from './ContextualMenuItem.types';
-
-const renderItemIcon = (props: IContextualMenuItemProps) => {
-  const { item, hasIcons, classNames } = props;
-
-  const { iconProps } = item;
-
-  if (!hasIcons) {
-    return null;
-  }
-
-  if (item.onRenderIcon) {
-    return item.onRenderIcon(props);
-  }
-
-  return <Icon {...iconProps} className={classNames.icon} />;
-};
-
-const renderCheckMarkIcon = ({ onCheckmarkClick, item, classNames }: IContextualMenuItemProps) => {
-  const isItemChecked = getIsChecked(item);
-  if (onCheckmarkClick) {
-    const onClick = (e: React.MouseEvent<HTMLElement>) => onCheckmarkClick(item, e);
-
-    return <Icon iconName={isItemChecked ? 'CheckMark' : ''} className={classNames.checkmarkIcon} onClick={onClick} />;
-  }
-  return null;
-};
-
-const renderItemName = ({ item, classNames }: IContextualMenuItemProps) => {
-  if (item.text || item.name) {
-    return <span className={classNames.label}>{item.text || item.name}</span>;
-  }
-  return null;
-};
-
-const renderSecondaryText = ({ item, classNames }: IContextualMenuItemProps) => {
-  if (item.secondaryText) {
-    return <span className={classNames.secondaryText}>{item.secondaryText}</span>;
-  }
-  return null;
-};
-
-const renderSubMenuIcon = ({ item, classNames }: IContextualMenuItemProps) => {
-  if (hasSubmenu(item)) {
-    return (
-      <Icon
-        iconName={getRTL() ? 'ChevronLeft' : 'ChevronRight'}
-        {...item.submenuIconProps}
-        className={classNames.subMenuIcon}
-      />
-    );
-  }
-  return null;
-};
+import { IContextualMenuItemProps, IContextualMenuItemStyles } from './ContextualMenuItem.types';
+import { IProcessedStyleSet } from '@uifabric/styling';
 
 export class ContextualMenuItemBase extends BaseComponent<IContextualMenuItemProps, {}> {
+  private _classNames: IProcessedStyleSet<IContextualMenuItemStyles>;
+
   public render() {
     const { item, classNames } = this.props;
 
+    this._classNames = classNames;
+
     return (
-      <div className={item.split ? classNames.linkContentMenu : classNames.linkContent}>
-        {renderCheckMarkIcon(this.props)}
-        {renderItemIcon(this.props)}
-        {renderItemName(this.props)}
-        {renderSecondaryText(this.props)}
-        {renderSubMenuIcon(this.props)}
+      <div className={item.split ? this._classNames.linkContentMenu : this._classNames.linkContent}>
+        {this._renderCheckMarkIcon()}
+        {this._renderItemIcon()}
+        {this._renderItemName()}
+        {this._renderSecondaryText()}
+        {this._renderSubMenuIcon()}
       </div>
     );
   }
@@ -95,4 +47,73 @@ export class ContextualMenuItemBase extends BaseComponent<IContextualMenuItemPro
       dismissMenu(undefined /* ev */, dismissAll);
     }
   };
+
+  private _renderCheckMarkIcon() {
+    const { onCheckmarkClick, item } = this.props;
+
+    const isItemChecked = getIsChecked(item);
+    if (onCheckmarkClick) {
+      const onClick = (e: React.MouseEvent<HTMLElement>) => onCheckmarkClick(item, e);
+
+      return (
+        <Icon
+          iconName={isItemChecked ? 'CheckMark' : ''}
+          className={this._classNames.checkmarkIcon}
+          onClick={onClick}
+        />
+      );
+    }
+    return null;
+  }
+
+  private _renderItemIcon() {
+    const { item, hasIcons } = this.props;
+
+    const { iconProps } = item;
+
+    if (!hasIcons) {
+      return null;
+    }
+
+    if (item.onRenderIcon) {
+      return item.onRenderIcon(this.props);
+    }
+
+    return <Icon {...iconProps} className={this._classNames.icon} />;
+  }
+
+  private _renderItemName() {
+    const { item } = this.props;
+
+    if (item.text || item.name) {
+      return <span className={this._classNames.label}>{item.text || item.name}</span>;
+    }
+
+    return null;
+  }
+
+  private _renderSecondaryText() {
+    const { item } = this.props;
+
+    if (item.secondaryText) {
+      return <span className={this._classNames.secondaryText}>{item.secondaryText}</span>;
+    }
+
+    return null;
+  }
+
+  private _renderSubMenuIcon() {
+    const { item } = this.props;
+
+    if (hasSubmenu(item)) {
+      return (
+        <Icon
+          iconName={getRTL() ? 'ChevronLeft' : 'ChevronRight'}
+          {...item.submenuIconProps}
+          className={this._classNames.subMenuIcon}
+        />
+      );
+    }
+    return null;
+  }
 }
