+++ packages/office-ui-fabric-react/src/components/ContextualMenu/ContextualMenuItem.test.tsx
@@ -3,16 +3,16 @@ jest.mock('../../utilities/contextualMenu');
 import * as React from 'react';
 import { shallow, ShallowWrapper } from 'enzyme';
 import { ContextualMenuItemBase } from './ContextualMenuItem.base';
-import { IContextualMenuItemProps } from './ContextualMenuItem.types';
-import { IContextualMenuItem } from './ContextualMenu.types';
-import { IMenuItemClassNames } from './ContextualMenu.classNames';
+import { IContextualMenuItemProps, IContextualMenuItem } from './ContextualMenuItem.types';
+import { IContextualMenuItemStyles } from './ContextualMenuItem.types';
 import { hasSubmenu } from '../../utilities/contextualMenu/index';
+import { IProcessedStyleSet } from '@uifabric/styling/lib/MergeStyles';
 
 describe('ContextMenuItemChildren', () => {
   describe('when a checkmark icon', () => {
     let onCheckmarkClick: jest.Mock;
     let menuItem: IContextualMenuItem;
-    let menuClassNames: IMenuItemClassNames;
+    let menuClassNames: IProcessedStyleSet<IContextualMenuItemStyles>;
     let wrapper: ShallowWrapper<IContextualMenuItemProps, {}>;
 
     beforeEach(() => {
@@ -51,7 +51,7 @@ describe('ContextMenuItemChildren', () => {
   describe('when it has icons', () => {
     describe('when it has iconProps', () => {
       let menuItem: IContextualMenuItem;
-      let menuClassNames: IMenuItemClassNames;
+      let menuClassNames: IProcessedStyleSet<IContextualMenuItemStyles>;
       let wrapper: ShallowWrapper<IContextualMenuItemProps, {}>;
 
       beforeEach(() => {
@@ -70,7 +70,7 @@ describe('ContextMenuItemChildren', () => {
 
     describe('when it doesnt have iconProps', () => {
       let menuItem: IContextualMenuItem;
-      let menuClassNames: IMenuItemClassNames;
+      let menuClassNames: IProcessedStyleSet<IContextualMenuItemStyles>;
       let wrapper: ShallowWrapper<IContextualMenuItemProps, {}>;
 
       beforeEach(() => {
@@ -90,7 +90,7 @@ describe('ContextMenuItemChildren', () => {
 
   describe('when it has a sub menu', () => {
     let menuItem: IContextualMenuItem;
-    let menuClassNames: IMenuItemClassNames;
+    let menuClassNames: IProcessedStyleSet<IContextualMenuItemStyles>;
     let wrapper: ShallowWrapper<IContextualMenuItemProps, {}>;
 
     beforeEach(() => {
@@ -109,7 +109,7 @@ describe('ContextMenuItemChildren', () => {
   });
 });
 
-function getMenuItemClassNames(): IMenuItemClassNames {
+function getMenuItemClassNames(): IProcessedStyleSet<IContextualMenuItemStyles> {
   return {
     item: 'item',
     divider: '---',
@@ -123,6 +123,9 @@ function getMenuItemClassNames(): IMenuItemClassNames {
     splitContainer: 'splitContainer',
     splitPrimary: 'splitPrimary',
     splitMenu: 'splitMenu',
-    linkContentMenu: 'linkContentMenu'
+    linkContentMenu: 'linkContentMenu',
+    anchorLink: 'anchorLink',
+    iconColor: 'iconColor',
+    subComponentStyles: {}
   };
 }
