+++ packages/office-ui-fabric-react/src/components/ContextualMenu/ContextualMenuItem.types.ts
@@ -1,7 +1,231 @@
-import { IContextualMenuItem } from './ContextualMenu.types';
-import { IMenuItemClassNames } from './ContextualMenu.classNames';
-import { IStyle, ITheme } from '../../Styling';
-import { IRefObject, IStyleFunction } from '../../Utilities';
+import { ITheme, IStyle, IProcessedStyleSet } from '../../Styling';
+import { IRefObject, IRenderFunction } from '../../Utilities';
+import { IButtonStyles } from 'office-ui-fabric-react/lib/components/Button';
+import { IVerticalDividerClassNames } from 'office-ui-fabric-react/lib/Divider';
+import { IStyleFunctionOrObject } from '@uifabric/utilities';
+import { ContextualMenuItemType, IContextualMenuProps, IContextualMenuSection } from './ContextualMenu.types';
+import { IIconProps } from '../Icon/Icon.types';
+import { IKeytipProps } from '../Keytip/Keytip.types';
+
+export interface IContextualMenuItem {
+  /**
+   * Optional callback to access the IContextualMenuRenderItem interface. This will get passed down to ContextualMenuItem.
+   */
+  componentRef?: IRefObject<IContextualMenuRenderItem>;
+
+  /**
+   * Unique id to identify the item
+   */
+  key: string;
+
+  /**
+   * Text description for the menu item to display
+   */
+  text?: string;
+
+  /**
+   * Seconday description for the menu item to display
+   */
+  secondaryText?: string;
+
+  itemType?: ContextualMenuItemType;
+
+  /**
+   * Props that go to the IconComponent
+   */
+  iconProps?: IIconProps;
+
+  /**
+   * Custom render function for the menu item icon
+   */
+  onRenderIcon?: IRenderFunction<IContextualMenuItemProps>;
+
+  /**
+   * Props that go to the IconComponent used for the chevron.
+   */
+  submenuIconProps?: IIconProps;
+
+  /**
+   * Whether the menu item is disabled
+   * @defaultvalue false
+   */
+  disabled?: boolean;
+
+  /**
+   * If the menu item is a split button, this prop disables purely the primary action of the button.
+   * @defaultvalue false
+   */
+  primaryDisabled?: boolean;
+
+  /**
+   * [TODO] Not Yet Implemented
+   */
+  shortCut?: string;
+
+  /**
+   * Whether or not this menu item can be checked
+   * @defaultvalue false
+   */
+  canCheck?: boolean;
+
+  /**
+   * Whether or not this menu item is currently checked.
+   * @defaultvalue false
+   */
+  checked?: boolean;
+
+  /**
+   * Whether or not this menu item is a splitButton.
+   * @defaultvalue false
+   */
+  split?: boolean;
+
+  /**
+   * Any custom data the developer wishes to associate with the menu item.
+   */
+  data?: any;
+
+  /**
+   * Callback issued when the menu item is invoked. If ev.preventDefault() is called in onClick, click will not close menu.
+   * Returning true will dismiss the menu even if ev.preventDefault() was called.
+   */
+  onClick?: (
+    ev?: React.MouseEvent<HTMLElement> | React.KeyboardEvent<HTMLElement>,
+    item?: IContextualMenuItem
+  ) => boolean | void;
+
+  /**
+   * An optional URL to navigate to upon selection
+   */
+  href?: string;
+
+  /**
+   * An optional target when using href
+   */
+  target?: string;
+
+  /**
+   * An optional rel when using href. If target is _blank rel is defaulted to a value to prevent clickjacking.
+   */
+  rel?: string;
+
+  /**
+   * Properties to apply to a submenu to this item.
+   * The ContextualMenu will provide default values for 'target', 'onDismiss', 'isSubMenu',
+   *  'id', 'shouldFocusOnMount', 'directionalHint', 'className', and 'gapSpace', all of which
+   *  can be overridden.
+   */
+  subMenuProps?: IContextualMenuProps;
+
+  /**
+   * Method to provide the classnames to style the individual items inside a menu. Default value is the getItemClassnames func
+   * defined in ContextualMenu.classnames.
+   * @default getItemClassNames
+   * @deprecated Use `styles` prop of `IContextualMenuItemProps` to leverage mergeStyle API.
+   */
+  getItemClassNames?: (
+    theme: ITheme,
+    disabled: boolean,
+    expanded: boolean,
+    checked: boolean,
+    isAnchorLink: boolean,
+    knownIcon: boolean,
+    itemClassName?: string,
+    dividerClassName?: string,
+    iconClassName?: string,
+    subMenuClassName?: string,
+    primaryDisabled?: boolean
+  ) => IProcessedStyleSet<IContextualMenuItemStyles>;
+
+  /**
+   * Method to provide the classnames to style the Vertical Divider of a split button inside a menu. Default value is the getVerticalDividerClassnames func defined in ContextualMenu.classnames
+   * @default getSplitButtonVerticalDividerClassNames
+   */
+  getSplitButtonVerticalDividerClassNames?: (theme: ITheme) => IVerticalDividerClassNames;
+
+  /**
+   *  Properties to apply to render this item as a section.
+   *  This prop is mutually exclusive with subMenuProps.
+   */
+  sectionProps?: IContextualMenuSection;
+
+  /**
+   * Additional css class to apply to the menu item
+   * @defaultvalue undefined
+   */
+  className?: string;
+
+  /**
+   * Additional styles to apply to the menu item
+   * @defaultvalue undefined
+   * @deprecated -- keco -- worth doing?
+   */
+  style?: React.CSSProperties;
+
+  /**
+   * Optional accessibility label (aria-label) attribute that will be stamped on to the element.
+   * If none is specified, the aria-label attribute will contain the item name
+   */
+  ariaLabel?: string;
+
+  /**
+   * Optional title for displaying text when hovering over an item.
+   */
+  title?: string;
+
+  /**
+   * Method to custom render this menu item.
+   * For keyboard accessibility, the top-level rendered item should be a focusable element
+   * (like an anchor or a button) or have the `data-is-focusable` property set to true.
+   *
+   * The function receives a function that can be called to dismiss the menu as a second argument.
+   *  This can be used to make sure that a custom menu item click dismisses the menu.
+   * @defaultvalue undefined
+   */
+  onRender?: (item: any, dismissMenu: (ev?: any, dismissAll?: boolean) => void) => React.ReactNode;
+
+  /**
+   * A function to be executed onMouseDown. This is executed before an onClick event and can
+   * be used to interrupt native on click events as well. The click event should still handle
+   * the commands. This should only be used in special cases when react and non-react are mixed.
+   */
+  onMouseDown?: (item: IContextualMenuItem, event: any) => void;
+
+  /**
+   * Optional override for the role attribute on the menu button. If one is not provided, it will
+   * have a value of menuitem or menuitemcheckbox.
+   */
+  role?: string;
+
+  /**
+   * When rendering a custom component that is passed in, the component might also be a list of
+   * elements. We want to keep track of the correct index our menu is using based off of
+   * the length of the custom list. It is up to the user to increment the count for their list.
+   */
+  customOnRenderListLength?: number;
+
+  /**
+   * Keytip for this contextual menu item
+   */
+  keytipProps?: IKeytipProps;
+
+  /**
+   * Any additional properties to use when custom rendering menu items.
+   */
+  [propertyName: string]: any;
+
+  /**
+   * This prop is no longer used. All contextual menu items are now focusable when disabled.
+   * @deprecated in 6.38.2 will be removed in 7.0.0
+   */
+  inactive?: boolean;
+
+  /**
+   * Text description for the menu item to display
+   * @deprecated Use `text` instead.
+   */
+  name?: string;
+}
 
 export interface IContextualMenuRenderItem {
   /**
@@ -30,7 +254,7 @@ export interface IContextualMenuItemProps extends React.HTMLAttributes<IContextu
   /**
    * Call to provide customized styling that will layer on top of the variant rules.
    */
-  styles?: IStyleFunction<IContextualMenuItemStyleProps, IContextualMenuItemStyles>;
+  styles?: IStyleFunctionOrObject<IContextualMenuItemStyleProps, IContextualMenuItemStyles>;
 
   /**
    * Theme provided by High-Order Component.
@@ -51,7 +275,7 @@ export interface IContextualMenuItemProps extends React.HTMLAttributes<IContextu
   /**
    * Classnames for different aspects of a menu item
    */
-  classNames: IMenuItemClassNames;
+  classNames: IProcessedStyleSet<IContextualMenuItemStyles>;
 
   /**
    * Index of the item
@@ -98,18 +322,37 @@ export interface IContextualMenuItemStyleProps {
   theme: ITheme;
 
   /**
-   * Accept custom classNames
+   * Additional CSS class(es) to apply to the ContextualMenuItem.
    */
   className?: string;
 
   // Insert ContextualMenuItem style props below
+  disabled: boolean;
+  expanded: boolean;
+  checked: boolean;
+  isAnchorLink: boolean;
+  knownIcon: boolean;
+  itemClassName?: string;
+  dividerClassName?: string;
+  iconClassName?: string;
+  subMenuClassName?: string;
+  primaryDisabled?: boolean;
 }
 
-export interface IContextualMenuItemStyles {
-  /**
-   * Style for the root element.
-   */
+export interface IContextualMenuItemStyles extends IButtonStyles {
+  item: IStyle;
+  divider: IStyle;
   root: IStyle;
-
-  // Insert ContextualMenuItem classNames below
+  linkContent: IStyle;
+  anchorLink: IStyle;
+  icon: IStyle;
+  iconColor: IStyle;
+  checkmarkIcon: IStyle;
+  subMenuIcon: IStyle;
+  label: IStyle;
+  secondaryText: IStyle;
+  splitContainer: IStyle;
+  splitPrimary: IStyle;
+  splitMenu: IStyle;
+  linkContentMenu: IStyle;
 }
